Update patch set 4
Patch Set 4:

There is an additional problem: some devices do not advertise themselves using the apropriate class and subclass, although they ar eregister-compatible (or close enough). This is the case for the SD card reader in my Thinkpad X220.

As a result, this approach chan't work, it's really up to the driver to make the decision and this cannot be hardcoded in the device manager according to PCI info.

Patch-set: 4