Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> That's the reason for introducing BMetaData, all those informations can be easily stored there. NOTE that in stagefright MetaData is used primarily for media headers.
> 
> I suggest to use BMetaData please introduce new keys where needed.

Yes, that would be better, even if it's a bit tricky to make sure we carry over all the information needed by ffmpeg and completely recreate the ffmpeg data structures from BMetaData so we can use the ffmpeg API without problems.

ffmpeg is not designed for complete decoupling of codec and format decapsulation as we are doing (and I'm not sure there is a clean way to do it, as there is a lot of variation on where data is stored in different media formats: in the frame metadata or in the stream one).

Patch-set: 3