Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-2
> > 
> > That's the reason for introducing BMetaData, all those informations can be easily stored there. NOTE that in stagefright MetaData is used primarily for media headers.
> > 
> > I suggest to use BMetaData please introduce new keys where needed.
> 
> Yes, that would be better, even if it's a bit tricky to make sure we carry over all the information needed by ffmpeg and completely recreate the ffmpeg data structures from BMetaData so we can use the ffmpeg API without problems.
> 
> ffmpeg is not designed for complete decoupling of codec and format decapsulation as we are doing (and I'm not sure there is a clean way to do it, as there is a lot of variation on where data is stored in different media formats: in the frame metadata or in the stream one).

Even with ffmpeg itself it's tricky to recreate everything when different media formats are involved. Their API is a minimum common denominator anyway. In general it's a problem that can't be solved completely using a generic API.

I see what you want to do, and it's very probable the only way is to pass some private stuff. But we still have the ability to add non-ffmpeg plugins, so we should try to export as much data as we can. Otherwise there's no point in having codec addons, right? We could just wrap it in a C++ friendly API and remove all this code in between that's painful to handle with.

This is not an easy situation, unfortunately, and even with BMetaData we can still have a keys for custom data, but we also want this data to be available to native apps when possible...

Patch-set: 3