Update patch set 4
Patch Set 4: Code-Review-2

> Patch Set 4: Code-Review-1
> sudo qemu-system-x86_64 --cpu host -smp 4 -k sv -m 1024 -enable-kvm -vga vmware -no-reboot -usb -usbdevice tablet -serial file:out.txt -cdrom generated/haiku-nightly-anyboot.iso -drive file=/dev/sdb,index=0,media=disk,format=raw -boot d

I am now able to reproduce this. The key is to run it as a CD, not as a disk. Your command showed that but I initially didn't think it would make a difference. It apparently does, as with this setup it's easily reproducible.

The Installer seems to end up with a lot of rpmalloc areas (~700MiB) until the system finally runs out of memory. Initially I suspected that the Installer was just leaking memory and that with the slight overhead of rpmalloc it would manage to exhaust the system. But I ran Installer through libroot_debug and the leak_analyzer and there does not seem to be an actual leak (i.e. at program termination there are no remaining allocations). Since this ran with the guarded heap, which has insane overhead for each allocation, it also does not seem like the Installer just generally uses a lot of memory. So it indeed looks like a problem specific to rpmalloc.

I'm currently investigating, but my spare time will soon run out, so not sure I'll get to the bottom of this.

Patch-set: 4
Reviewer: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2