Update patch set 4
Patch Set 4:

> I am now able to reproduce this. The key is to run it as a CD, not as a disk. Your command showed that but I initially didn't think it would make a difference. It apparently does, as with this setup it's easily reproducible.

Running as a CD increases the memory pressure because of the write_overlay, which will cache all writes to the CD in RAM. It may also change the relative speed of reading from the source and writing to the destination. I don't know if these are synchronized with each other in Installer, or if it will just read as much as possible from the source even if the destination is slower to write?

Patch-set: 4