Update patch set 3
Patch Set 3:

Honestly, I think the idea of using B_MEDIA_ENCODED_AUDIO and B_MEDIA_ENCODED_VIDEO is very buggy. The media_kit tries to give a common description of those, with an hilarious result. Leave apart the fact that 80% of those struct fields are left unused, and in BMediaFormat only those actually useful will be leaved, it makes the whole thing hard to deal with. Not to mention the multistream stuff.

ffmpeg and stagefright, as well as other APIs, just uses MIME types for that. So I think the next move is to do the same. We should have a system defined MIME for RAW Audio and Video, then all the other stuff should work using MIME types. Instead of using the media_type in BMediaFormat we just use a string MIME, and find the codecs that way. We can in the end just copy parts of ffmpeg, that should be helpful in making everything more transparent.

In my upcoming merge I'm not going to do that, but this is planned in future.

Patch-set: 3
Reviewer: Gerrit User 1000042 <1000042@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0