Update patch set 4
Patch Set 4:

It looks like the Installer case is a pathological case for how the defaults are set up. Due to the pattern of allocation in one thread and deallocation in the other, allocated memory is shifted into the thread cache of the latter, where it is never reused. The cache of the former then runs out and new memory is mapped only to be shifted into the other thread again.

The default configuration has a thread cache multiplier of 16, meaning that 16 times the release count needs to be reached before moving one release count of spans into the global cache. This guarantees that a lot of memory (15 times release count) stays in the span cache of the deallocating thread at all cases. Since the allocations here are large, the multiplier makes things add up quickly. I have not done exact measurements however.

In cases like these it would be nice if an allocation failure in one thread would try to use caches of other threads that may hold a large cache. Other than that a per thread heuristic of how likely memory allocations of a certain size are and then caching only a reasonable amount might work. Obviously this goes quite far into the implementation of rpmalloc and wouldn't really make sense as a Haiku specific patch.

Alternatively the defaults could simply be tuned. The cache multipliers are rather large, which is undoubtedly friendly to certain types of benchmarks and some real world use cases. For a general purpose OS allocator they are too extreme IMO. Reducing the multiplier to 8 already made the Installer work. Figuring out sane values will probably need some investigating.

Patch-set: 4