Update patch set 2
Patch Set 2:

AboutSystem is about the system and hardware, not about how the system is configured. So, e.g. the username or other details aren't listed there. Further, Network preferences are easily accessible from the Deskbar if someone wants to check the hostname, so it's not like we're sacrificing convenience. So I'd say we don't want the hostname in AboutSystem.

Patch-set: 2