Fix building test_app_server using GGC2
test_app_server needs to link against both the system's libbe.so
and libbe_test.so. In order to decouple the clashing classes, there
is an extra libhwinterface.so, containing only the HWInterface
definition, and libhwinterfaceimpl.so, which contains the actual
implementation using the system libbe's classes. This requires
undefined symbols to be allowed when building the shared libs
which are later assembled into the test_app_server.