Update patch set 1
Patch Set 1:

So, doing CheckSanity only on !=B_OK leads to shutdown stalling on bash or dstChk stuck on modal window. It was suggested this might be due to teams that got deleted before watching, so an early check was added after watching was setup. 

Further investigation seems to suggest that we can just do the removal of the team deleted so changed to that. I don't see any issues doing this, even without early check, but I only run the builds for minutes. This is as far as I can take this I think.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0