Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> On a side note: at the moment we don't include the hostid coreutils command line:
> https://github.com/haikuports/haikuports/blob/master/sys-apps/coreutils/patches/coreutils-8.29.patchset#L142

Yeah, the commandline manipulation of the hostid is not a part of the posix specification and I didn't think building it in is really warranted. Hence the sample code instead of including it in the patch set. 

It's not difficult for someone to set a random id in the file should they need gethostid(), the standard doesn't specify exactly how the hostid is to generated in any case.

gethostid() is somewhat of an anachronism, 32 bits seems remarkably tiny these days for a unique identifier, but it's right there in posix unfortunately.

Patch-set: 1
Reviewer: Gerrit User 1000091 <1000091@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0