Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Also, with this change in place, what is the error message that a user sees in DriveSetup? Will it be helpful as to what to do in order to be able to initialize a partition?

The patch makes use of the existing logic within DriveSetup, and the options which do not match the partition type are not made available for the user to select. No error message is displayed, because the user cannot select the disabled items.

Screenshot examples:

Selecting a Be File System partition: https://imgur.com/a/aekUsw7

Selecting the Intel Extended Partition: https://imgur.com/a/1hhrqJ9

Selecting an empty drive : https://imgur.com/a/llD6Vre

Patch-set: 2