Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi Rob, thanks for your work.
> 
> About the code itself: I wonder if the comparison "const char* == BSTring" works? I know that "BString == const char*" has the expected result. Also there seem to be a few parenthesis too many.

The comparison works. I'll remove some of the parentheses to match code style.
> 
> Now about the chane: I wonder whether this change won't also prevent a number of valid situations where the user wants to initialize the correct partition, but it doesn't have the supported type and the user also doesn't know how to change the type. After all, I have recommended for a long time that users change the partition table with some other, trusted tool, and then only initialize with BFS in DriveSetup. It seems likely that the partitions will not have the correct type. What is your thinking?

A minimal change like this covers most use cases, but no it does not not cover the case where someone has partitioned the drive to an incorrect type and doesn't know to repartition it to the correct type.

For a user facing gui, to minimize the sort of problems seen, it's tempting to imagine rolling partition creation and partition formatting into one. But this would be an enormous change to the way DriveSetup works.

Patch-set: 2
Label: Code-Review=0