Update patch set 3
Patch Set 3:

> Patch Set 2:

It would be possible that the logic could be reversed, and specifically disallow installing to the Intel Extended Partition container, but leave all other types as fair game.

It wouldn't prevent formatting within incorrect partition types and the problems seen in #14839, but would avoid catastrophic partition loss as in #9856.

Patch-set: 3
Label: Code-Review=0