Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 2:
> 
> It would be possible that the logic could be reversed, and specifically disallow installing to the Intel Extended Partition container, but leave all other types as fair game.
> 
> It wouldn't prevent formatting within incorrect partition types and the problems seen in #14839, but would avoid catastrophic partition loss as in #9856.

I am leaning towards preferring that. And at the same time, the partition type ID should be set to match the file system. Is this possible? My impression is that formatting an empty partition with BeFS would not fix #14839, unless it also sets the type, or am I missing something?

Patch-set: 3
Label: Code-Review=0