Update patch set 3
Patch Set 3:

I don't like this solution much from the UI point of view. The menu item becomes invisible or disabled. This does not provide any hint to the user as to why the action is not possible, and in this case, it is not obvious.

Moreover, the problem is easily solved by setting the partition type automatically when formatting. This change is indeed more involved, but it would make more sense for the user. The partition type is something a little internal, and it's better for us to just make sure it matches the actual filesystem, without letting the user worry about that.

As for erasing an intel extended partition with a filesystem, maybe an alert is due in that case. With the current DriveSetup design, this may be a little difficult to implement. Maybe check if the partition we're about to initialize has sub-partitions, and pop an alert in that case? I would not completely prevent it, because it is a valid use case, even if uncommon.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0