Update patch set 2
Patch Set 2:

> Also, seeing a lot of scroll bars where they were once not.

Yes, BScrollView does not auto-hide scrollbars when they aren't needed. I tried to hack that into DynamicScrollView, but that really can't work because it would need to resize the view in order for it to work correctly.

Prior to this commit, the scrollviews were kinda weird: you could use them, but if you ever resized the window larger, you couldn't make it smaller again. Now at least we are just using BScrollView so the behavior is consistent with the rest of the system.

Patch-set: 2