Update patch set 2
Patch Set 2:

> It's correct, but a lot uglier.  Anything we can do?

There are now scrollbars permanently attached to the views, yes, which is BScrollView's default behavior. Making them "go away" automatically would require some rather invasive changes, which we should probably do, but is really outside the scope of this patch. Otherwise it's nearly identical to the old look, but with much better layout and sizing.

Patch-set: 2
Label: Code-Review=0