Update patch set 2
Patch Set 2:

In the absence of an host id file, the glibc implementation uses the local IPv4 address converted to a 32bit int. This provides an unique identifier on the local network, at least, without user intervention.

Moreover, POSIX does not allow this function to raise an error, and as a result, setting errno is not nice (but apparently allowed by POSIX). Anyway, it means portable POSIX code cannot check errno to see if this function worked. Interestingly, checking that the return value is negative doesn't work either, because the function returns a 32 bit host id, meaning negative values are valid host IDs.

However, just getting the IP address is not correct either: it may change over time, for example if you are roaming various wifi networks.

And a final note, instead of opening the file everytime the function is called, one could cache the result in a static variable, making sure the host id does not change at least during the runtime of one application.

Well, all of this is probably overengineering stuff for a function that can't be used anyways...

Patch-set: 2