Update patch set 3
Patch Set 3:

(3 comments)

I think the destinction in the code between sector_size and block_size, and why one is used here and the other is used there, is not at all clear yet. And some methods which appear to take sector size, need to be renamed to make that clear.

Patch-set: 3