Update patch set 2
Patch Set 2:

I'm not sure what's unclear; we should keep image_id easily accessible inside crtn, even if we don't use it there right now. Other architectures with callee-cleanup-args as default preserve it automatically, and we don't need to move it onto the stack explicitly. So rather than removing this code (which was fixed multiple times for stack alignment), we can just leave it as-is in case we need it again. It's only 4 instructions anyway.

Patch-set: 2