Update patch set 2
Patch Set 2:

Hi,
After improving the tracing, things are clearer now.

So, the SDHCI bus gets to probe the generic and isa root devices, as well as the SMBus controller in my machine. But, it does not get to probe the SD controller.

This is expected, because currently we are expecting it to be called only when the PCI bus encounters an SD controller card (this is hardcoded in device_manager.cpp). However, in my case the controller does not advertise itself as SDHCI compatible (it has type 8 / subtype 80: generic, other system peripheral). Should I let the MMC bus scan this, too?

Patch-set: 2
Label: Code-Review=0