Update patch set 1
Patch Set 1: Code-Review-2

The Command Key bitmap is most likely fine, but choosing when to use it is tricky and this patch does not address that well enough. This change really only makes sense on a Mac Keyboard which has a Command key on it with the clover symbol. More discussion here https://dev.haiku-os.org/ticket/7967 and here https://www.freelists.org/post/haiku-development/Key-roles-vs-key-label-in-menus .

Unfortunately we may not be able to improve on the compromise we currently have in place.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2