Update patch set 1
Patch Set 1:

> More discussion here https://dev.haiku-os.org/ticket/7967 and here https://www.freelists.org/post/haiku-development/Key-roles-vs-key-label-in-menus .

Thanks for directing me to the historical discussion on this issue and the context for why the code I am looking at exists the way it does today. I saw you mention in the thread that you wanted the Command and Option functions for Mac keyboards to automatically be mapped correctly, but as of today that has not been done yet, so a manual keymap adjustment is required. Fine with me, but as it stands, on a Mac or similar keyboard, with the keymap adjusted to match, the menu labels for Command/"ALT" and Option are reversed, and I think this is a very bad look, especially when there is no "Alt" label on any key on recent Apple keyboards, and the one being presented as "OPT" is not the one so labeled. This is exactly the issue it sounds like you were trying to avoid in the original discussion. I am used to fighting with different combinations of keyboards and platforms and this is disruptive even for me; it must be terrible for a new Haiku user installing it for the first time on their MacBook. That's just the case that something should be done, though. Perhaps more important in this context are thoughts about this specific patch.

> The Command Key bitmap is most likely fine, but choosing when to use it is tricky and this patch does not address that well enough. This change really only makes sense on a Mac Keyboard which has a Command key on it with the clover symbol. 

My outlook is more optimistic. I have no data on this, but I think we can assume that most keyboards with modifiers laid out this way are Macintosh keyboards, which map perfectly to the internal key roles and to this patch, once the keymap is adjusted. Second most common are probably Happy Hacking Keyboards, which also match because they contain the same labels as the Mac keyboards. Sun keyboards are probably much more rare, and a less perfect match because the Alt/Opt key does not actually have "Opt" printed on it, but still a better match than the current behavior, and I don't think it would be overly ambitious to assume that somebody using a Sun keyboard on a Haiku machine is aware of the various common associations of Command/Option/Meta/Alt/etc. Giving the "diamond" key on such a keyboard the actual role label "Command" is as good a match as any, certainly better than Alt and especially Option. A much more obscure example would be a NeXT keyboard connected through an adapter, which would also be better-served by this patch than the status quo.

I cannot think of any existing keyboard with any likelihood of being connected to a Haiku machine that would be adversely impacted by this patch's representation, although I would be very interested in knowing about any if they do.

> Unfortunately we may not be able to improve on the compromise we currently have in place.

I agree that this is not the absolute best way to do it, but for the above-stated reasons I consider it an improvement that does away with a serious visual issue on a fairly common series of hardware as well as some more obscure ones, without stepping on the toes of anything else.

Patch-set: 1