Update patch set 1
Patch Set 1:

Skipping over the backstory again we came to the conclusion that the way to move forward here is to map the items to their "key roles" aka B_COMMAND_KEY, B_OPTION_KEY, etc. so that MenuItemShift() always returns sMenuItemShift, MenuItemControl() always returns sMenuItemControl, MenuItemOption() always returns sMenuItemOption and MenuItemCommand() always returns sMenuItemCommand. The menu never changes and you adjust your keymap to match what you want. Command and Option are reversed on Mac keyboards compared to Alt and the Windows Key on a PC keyboard. This means abandoning the idea of putting Alt in the menu since we don't have an Alt key role, just command. It's either that, or we leave things the way they are now.

Patch-set: 1