Update patch set 1
Patch Set 1:

In that case I would expect the default bitmap to be CMD rather than ALT, but even then that still raises the problem of the bitmaps being wrong when the keymap is adjusted for a Mac keyboard because the code is looking for specific keycodes, not just the role, to decide which bitmap to provide. If it is going to factor in keycodes, should some accomodation not be made for that? Or is the solution to stop checking keycodes entirely and provide a bitmap based on the role alone? I like that idea personally, but the linked discussion happened seven years ago and nobody threw the switch, so I thought maybe there was some reason why people still wanted the original behavior and tried to work with it.

Patch-set: 1