Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> In that case I would expect the default bitmap to be CMD rather than ALT

Yes should be command. Alt does not represent a key role, it represents a key and we want to show roles in the menu because it is impossible to show keys accurately.

> but even then that still raises the problem of the bitmaps being wrong when the keymap is adjusted for a Mac keyboard because the code is looking for specific keycodes, not just the role, to decide which bitmap to provide. If it is going to factor in keycodes, should some accomodation not be made for that? Or is the solution to stop checking keycodes entirely and provide a bitmap based on the role alone? I like that idea personally, but the linked discussion happened seven years ago and nobody threw the switch, so I thought maybe there was some reason why people still wanted the original behavior and tried to work with it.

I think you have it perfectly. The discussion 7 years ago concluded that we go with key roles alone with no consideration of key map settings. But some still prefer the current situation, and no action has been taken until you came along and restarted the discussion with your patch. IF we could detect the keyboard and determine what kind it was we could provide more accurate glyphs, until then we need to determine what changes are needed here based on limited info we have.

The assumption of a Windows style keyboard with Windows key and Alt is expected and Mac, Happy Hacking and other keyboards are the exception. We can put that Cmd = Alt in the user guide but people won't read it. Maybe we could make a setting in keymap prefs and then even though we can't auto-detect the keyboard type you'll be able to set it there to get the right menu glyphs for your keyboard.

Patch-set: 1