Update patch set 1
Patch Set 1:

> I think you have it perfectly. The discussion 7 years ago concluded that we go with key roles alone with no consideration of key map settings. But some still prefer the current situation, and no action has been taken until you came along and restarted the discussion with your patch. IF we could detect the keyboard and determine what kind it was we could provide more accurate glyphs, until then we need to determine what changes are needed here based on limited info we have.
> 

If we lack sufficient consensus to go with roles alone, I personally I feel like Command/Option presentation in the fashion of a Mac keyboard is a reasonably good cover for everything that is not a Windows-style keyboard, and we can detect when it is not by the way the keymap is configured, using the same type of keycode checks already in place (this patch). By that reasoning, it is a binary choice and there is actually very little we have to detect to get a reasonably good match. Going with roles only is better, but if we don't want to I think it is an improvement. Unless I am mistaken, in terms of all types of the keyboards that can connect to a modern PC, Windows-style is the exception and everything else works about the same in terms of these two keys.

> Maybe we could make a setting in keymap prefs and then even though we can't auto-detect the keyboard type you'll be able to set it there to get the right menu glyphs for your keyboard.

This I also consider reasonable.

Patch-set: 1
Reviewer: Gerrit User 1000099 <1000099@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0