Update patch set 1
Patch Set 1:

I would be very unhappy with a setting for this. It is a minor detail the user should not care about. If we can autodetect and display the right bitmap, that's great, however if it needs manual setup, I don't think it's worth it.

I'm fine with whatever we display in the menu (as I mentionned on IRC, it will never fully match my keyboard due to the way we mix alt and windows keys to implement AltGr). But I'm not fine with a setting to choose what to display.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0