Update patch set 4
Patch Set 4: Code-Review-2

(11 comments)

I still think you should store the autostart setting in a single place, in the application. You now have 3 copies of the setting, which complicates the code a lot to make sure they are all kept in sync, and makes little sense because this really is a setting for the whole application.

By routing the messages to the right place, you can make the menu item send a message to the application directly, so there is no need to do things in the window and launch button classes.

Patch-set: 4
Label: Code-Review=-2