Update patch set 1
Patch Set 1:

> Of course I would still prefer to have an output parameter of type uint32 for the masks, so we can use unsigned integers here. The return value should be a status_t.

Except most of these functions simply take the bitfield, run a bitwise-AND on it, and then return the value. This would be an implicit unsigned-to-signed conversion, and thus throw a lot of compiler warnings without explicit casts. So keeping it all as one signedness makes much more sense here.

Patch-set: 1