Update patch set 1
Patch Set 1:

> You misunderstood: The return value should be a status_t and not be abused to return a mask.

Ah, now I understand.

We need to return a mask from these functions in order to determine what events are pending. Probably the return type should just be int32 then, so that positive values are the mask, and negative ones are the errors.

That's not "abusing" anything, as we already have that same behavior for create_area, etc.

Patch-set: 1