Update patch set 6
Patch Set 6:

(1 comment)

Ok, one last thing I though about this morning.

If you have multiple Launchbox windows, and you change the setting in one, the others will not be updated and their menus will show the wrong value for the setting.

We can fix this by implementing MenusBeginning() in the window, and refreshing the state of the menu item right before the menu opens.

Patch-set: 6