BeDecorator Draw outer left close button bevel to the edge
... just like BeOS R5 did.

In 31234326741231bb8d36456ce9dbdccec7ce0e3b (hrev52122) I wrote:

"Undraw the bottom left and top right corners of bitmaps in the
unpressed state. BeOS did not do this (usually), but it should
have as it closes the ring."

This was a bad idea, draw like BeOS R5 instead.