Update patch set 1
Patch Set 1: Code-Review-1

Hi,

The applications in the src/tests directory are meant to test only a specific part of the code at a time. So they should focus just on that part.

This application is about testing menus, so I think it should not be localized. Let's saay someday I run it and find no text in the menus. Is it a locale kit problem? Is it an actual menu problem? I wouldn't know.

Patch-set: 1
Label: Code-Review=-1