Update patch set 1
Patch Set 1:

> B_CONSTANT_ALPHA

sorry... I mean B_ALPHA_OVERLAY of course. COMPOSITE should only be needed when the target is not fully opaque. If this change fixes the problem for real, it should mean that the COMPOSITE calculation doesn't have the same calculation error as the OVERLAY calculation. Nevertheless, the problem should be fixed in the calculation, not by using a different calculation.

Patch-set: 1