Update patch set 1
Patch Set 1:

I will have a look at those headers. 

This patch does stop the colour change.

Prior to the patch it was changing RGB 51,102,152 to RGB 50,101,151 in the squarish region immediately around the icon. (I'm impressed that the original reporter even noticed that).

With the patch applied it remains RGB 51,102,152.

Patch-set: 1