Update patch set 1
Patch Set 1:

I hope you find your way around there. The basic idea is this: The AGG rendering backend takes a class called "pixel format", which must define a number of methods. (It is used as an input to a number of templated rendering classes.) This interface is implemented in PixelFormat.h and PixelFormat.cpp (I believe). To be able to dynamically switch drawing modes, this class has a number of function pointers. Since many drawing modes would still duplicate code a lot, there is DrawingMode.h with those macros. I would first search the error in those macros.

Patch-set: 1