Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I hope you find your way around there. The basic idea is this: The AGG rendering backend takes a class called "pixel format", which must define a number of methods. (It is used as an input to a number of templated rendering classes.) This interface is implemented in PixelFormat.h and PixelFormat.cpp (I believe). To be able to dynamically switch drawing modes, this class has a number of function pointers. Since many drawing modes would still duplicate code a lot, there is DrawingMode.h with those macros. I would first search the error in those macros.

Looks to be a rounding error of sorts.

x = ( pixel(x) * 255 + pixel'(x) << 8  ) >> 8

Effectively : (pixel x * 255 + pixel'(x) * 256 ) / 256
So the 

I can see a workaround in the COMPOSITE sections, where a specific test is added for alpha of zero. 

(Although I assume from your earlier statement that there are going to regions of low but non-zero alpha in the icons, so that workaround shouldn't solve that problem in that case)

Patch-set: 1