kernel/elf: fix build error on sparc
Casting char* to larger pointer types is not allowed on sparc. Use an
union to perform the conversion here, as we know the elf data makes
sense and is properly aligned (otherwise we will get SIGBUS at runtime).

I think this is also more readable than playing with casted pointers
everywhere.