Add BytePointer to help with pointer conversions
This behaves mostly like a pointer, but pointer maths works in bytes,
not the native object size.

This allows to generate misaligned pointers, so use with caution. It is
not possible to generate such pointers by casting on sparc, the compiler
rejects them because the architecture doesn't support unaligned access
(and would generate a SIGBUS).

This can be used when creating the address of an object from a pointer +
offset. We assume the caller knows what he is doing and will make sure
the alignment is correct. We could add an explicit alignment check when
built in debug mode, maybe?

Change-Id: I6a8681a398345f0c7d419a2cfe7244d972ffa62f