Update patch set 2
Patch Set 2:

From my initial experience with the sparc port, we will have to use this a lot, basically everytime we want to do pointer arithmetics in byte and then convert to a type with larger alignment.

So, it is important to get the implementation and design of this right. Please review carefully, and use the later commits submitted with this to get an idea how it looks like in practise.

Also, none of this was tested on x86, we must do that before merging, of course.

Patch-set: 2