Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Question: how does everyone else on sparc if not using such a class?


It turns out I was wrong, in most cases they just allow misaligned accesses, and catch them with a trap handler to perform a slow access instead. This class does not fix the problem with misaligned accesses anyway.

We can keep it still, if it makes the code more readable (not decided on this).

Patch-set: 2