Update patch set 1
Patch Set 1:

Spacing between the button and the textfield should be smaller. Maybe no space at all, or maybe same spacing as between the spinner textfield and buttons.

I think the icon feels not very Haiku-ish, but I'm not sure why. Maybe because an icon alone in a button is very unusual in Haiku (we have icons in toolbars but they don't have a button frame). Or maybe I'm just not used to these flat icons yet.

Also I wonder if a folder icon is appropriate here. What about a magnifying glass, maybe?

Patch-set: 1