Update patch set 1
Patch Set 1:

Thinking about this during dinner, there is an usability problem too.

The button is too small.

No one will be editing the path directly in the textfield. The button is the main way to interact with this. Yet, it is now a small 16x16px button (and I did not even check with smaller font sizes).

So, how do we usually do this? Backgrounds uses a BOptionPopUp instead. I think that's a nice idea. It does not allow direct editing of the path (a good thing, as only valid paths can be set then), and it allows to list a few common directories to pick from immediately, or open a file panel to select something else).

Debugger uses a similar solution, although currently not as a BOptionPopUp (but I think it should do that).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0