Update patch set 1
Patch Set 1:

> Or maybe I'm just not used to these flat icons yet.

Maybe... I quite like it. It is small (the button with a 13px is 26x26px), but nice and unobtrusive.
 
> Also I wonder if a folder icon is appropriate here. What about a magnifying glass, maybe?

The folder icon is normally used to open a file/open dialog, see Pe, Koder and most cross platform apps. A magnifying glass would suggest a search function like a query window.

Patch-set: 1