Update patch set 2
Patch Set 2:

GCC documentation explicitly says using unions this way works, for example here:
http://gcc.gnu.org/bugs/#nonbugs

It is indeed a gcc extension, but no special flags are needed.

I'm ok with a cast, it will be a little less readable but at least it is only in one place.

As for standards, I think the whole idea of casting to char, doing pointer math and casting back is not really allowed, but I don't think it's going to create any problems either?

Patch-set: 2