Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi,
> 
> The file you modified is meant to document, exclusively, how to contribute to the documentation in docs/user. Parts of what you added are off-topic here.
> 
> Some of it could go into docs/user in other files. This is documentation for users of the API, that is, people writing software for Haiku but not directly contributing to Haiku sources.
> 
> Other parts of what you added are targetted to Haiku contributors or people who want to contribute to Haiku, these should go to docs/develop or maybe to the main website (files for the website are hosted separately at github.com/haiku/website).

Hi 

In documentation part or section I modified the file in a way that it provides a series of steps or links to help contributors to contribute to haiku. 

Should I again modify it? or it is going to get merged in some other file in doc/user or in this file only.

Patch-set: 1
Reviewer: Gerrit User 1000106 <1000106@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0